feat: Refactor webhook repository
